+++ docs/merge_docs.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+sed -i '$a\\n' ../demo/README.md
+
 sed -i 's/(\/tools\/data\/activitynet\/preparing_activitynet.md/(#activitynet/g' supported_datasets.md
 sed -i 's/(\/tools\/data\/kinetics\/preparing_kinetics.md/(#kinetics/g' supported_datasets.md
 sed -i 's/(\/tools\/data\/mit\/preparing_mit.md/(#moments-in-time/g' supported_datasets.md
@@ -18,11 +20,13 @@ sed -i 's/(\/tools\/data\/ava\/preparing_ava.md/(#ava/g' supported_datasets.md
 cat  ../configs/localization/*/*.md > localization_models.md
 cat  ../configs/recognition/*/*.md > recognition_models.md
 cat  ../tools/data/*/*.md > prepare_data.md
+cat  ../demo/README.md > demo.md
 
 sed -i 's/#/##&/' localization_models.md
 sed -i 's/#/##&/' recognition_models.md
 sed -i 's/md###t/html#t/g' localization_models.md
 sed -i 's/md###t/html#t/g' recognition_models.md
+sed -i "s/md###t/html#t/g" demo.md
 
 sed -i 's/# Preparing/# /g' prepare_data.md
 sed -i 's/#/##&/' prepare_data.md
@@ -45,3 +49,4 @@ sed -i 's/](\/docs\//](/g' ./tutorials/*.md
 sed -i 's=](/=](https://github.com/open-mmlab/mmaction2/tree/master/=g' ./tutorials/*.md
 sed -i 's/](\/docs\//](/g' supported_datasets.md
 sed -i 's=](/=](https://github.com/open-mmlab/mmaction2/tree/master/=g' supported_datasets.md
+sed -i 's=](/=](https://github.com/open-mmlab/mmaction2/tree/master/=g' demo.md