+++ docs/changelog.md
@@ -7,6 +7,7 @@
 **New Features**
 
 **Improvements**
+- Add random seed for building filelists ([#323](https://github.com/open-mmlab/mmaction2/pull/323))
 
 **Bug Fixes**
 - Fix a bug in BaseDataset when `data_prefix` is None ([#314](https://github.com/open-mmlab/mmaction2/pull/314))